requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
carehare==1.0.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
daphne==4.1.2
defusedxml==0.7.1
Django==5.0
dnspython==2.7.0
eventlet==0.38.2
filelock==3.16.1
flower==2.0.1
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
msgpack==1.1.0
packaging==24.2
pamqp==3.3.0
parsel==1.9.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
Protego==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-memcached==1.62
python-scrapyd-api==2.1.2
pytz==2024.2
queuelib==1.7.0
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.2
tldextract==5.1.3
tornado==6.4.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.2
